Resolve 1629 problem

+++ src/pr_1629.rs
@@ -3,13 +3,13 @@ use std::io;
 pub fn pr_1629() {
     let mut buffer = String::new();
     io::stdin().read_line(&mut buffer).unwrap();
-    let num_list:Vec<u32> = buffer.split_ascii_whitespace().map(|x| x.trim().parse::<u32>().unwrap()).collect();
+    let num_list:Vec<u64> = buffer.split_ascii_whitespace().map(|x| x.trim().parse::<u64>().unwrap()).collect();
     let (a, b, c) = (num_list[0], num_list[1], num_list[2]);
 
     println!("{}", go(a, b, c));
 }
 
-fn go(a:u32, b:u32, c:u32) -> u32 {
+fn go(a:u64, b:u64, c:u64) -> u64 {
     println!("{}", a%c);
     if b == 1 { return a % c; }
